Force all per-node query response handling onto a single thread
Same reasoning as for field_caps in https://github.com/elastic/elasticsearch/pull/120863,
no need to have multiple threads contending the same mutex(s) when the heavy lifting
step in handling the results is sequential anyway.
